feat(character): returned file names
